Trail correction
After merging and testing the ball some more, I noticed the trail wasn't working. So fixed that bug. Also I made the walls of the stage WorldBoundaryShape2D instead of RectangleShape2D. It helps the collision and makes the walls infinitely tall.

Also added the p2smash to the input map so the debugger would stop shouting at me.