Points and Win Condition
Added the ability to win the game. Also added ball resetting after a point is scored. It still needs to run initial setup when the stage first starts to decide who gets to serve first. It also needs to actually do the graphics stuff, and I need to adjust the ball so it actually goes over the net when you hit it, or we need to adjust the point system itself.